+++ src/main/java/org/entity/TestEntity.java
@@ -15,7 +15,4 @@ public class TestEntity {
   public Answer getAnswer() {
     return answer;
   }
-  public void setAnswer(Answer answer) {
-    this.answer = answer;
-  }
 }
